+++ app/src/main/java/su/sres/securesms/profiles/ProfileMediaConstraints.java
@@ -17,27 +17,27 @@ public class ProfileMediaConstraints extends MediaConstraints {
   }
 
   @Override
-  public int getImageMaxSize(Context context) {
+  public long getImageMaxSize(Context context) {
     return 5 * 1024 * 1024;
   }
 
   @Override
-  public int getGifMaxSize(Context context) {
+  public long getGifMaxSize(Context context) {
     return 0;
   }
 
   @Override
-  public int getVideoMaxSize(Context context) {
+  public long getVideoMaxSize(Context context) {
     return 0;
   }
 
   @Override
-  public int getAudioMaxSize(Context context) {
+  public long getAudioMaxSize(Context context) {
     return 0;
   }
 
   @Override
-  public int getDocumentMaxSize(Context context) {
+  public long getDocumentMaxSize(Context context) {
     return 0;
   }
 }
